moving __skeleton__ into plugins
