--- __skeleton__/README.md
@@ -1,4 +1,0 @@
-# {{PLUGIN_NAME}}
-
-This is a generated file! This file should be updated with any specific instructions on how to install or use your plugin.
-
