--- __skeleton__/vite.config.ts
@@ -1,4 +1,0 @@
-import { defineConfig } from '@fiftyone/plugin-build'
-
-export default defineConfig(__dirname)
-
