+++ plugins/__skeleton__/package.json
@@ -1,0 +1,21 @@
+{
+    "name": "{{PLUGIN_NAME}}",
+    "version": "1.0.0",
+    "main": "src/index.ts",
+    "fiftyone": {
+        "script": "dist/index.umd.js"
+    },
+    "scripts": {
+        "dev": "nodemon --watch ./src --ext js,jsx,ts,tsx --exec 'yarn build'",
+        "build": "vite build",
+        "preview": "vite preview"
+    },
+    "dependencies": {
+        "react": "^18.2.0",
+        "@fiftyone/plugin-build": "*"
+    },
+    "devDependencies": {
+        "nodemon": "^2.0.19",
+        "vite": "^3.2.4"
+    }
+}
