+++ plugins/__skeleton__/src/Plugin.tsx
@@ -1,0 +1,11 @@
+import * as fos from "@fiftyone/state";
+import { useRecoilValue } from "recoil";
+
+export function Plugin() {
+  const dataset = useRecoilValue(fos.dataset);
+  return (
+    <h1>
+      You are viewing the <strong>{dataset.name}</strong> dataset.
+    </h1>
+  );
+}
