+++ plugins/__skeleton__/vite.config.ts
@@ -1,0 +1,3 @@
+import { defineConfig } from "@fiftyone/plugin-build";
+
+export default defineConfig(__dirname);
